[FLINK-18905][hotfix][task] Move output and collector helper classes out of OperatorChain
