+++ flink-streaming-java/src/main/java/org/apache/flink/streaming/runtime/tasks/ChainingOutput.java
@@ -1,0 +1,143 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.flink.streaming.runtime.tasks;
+
+import org.apache.flink.metrics.Counter;
+import org.apache.flink.metrics.Gauge;
+import org.apache.flink.metrics.SimpleCounter;
+import org.apache.flink.runtime.metrics.groups.OperatorIOMetricGroup;
+import org.apache.flink.runtime.metrics.groups.OperatorMetricGroup;
+import org.apache.flink.streaming.api.operators.OneInputStreamOperator;
+import org.apache.flink.streaming.api.watermark.Watermark;
+import org.apache.flink.streaming.runtime.metrics.WatermarkGauge;
+import org.apache.flink.streaming.runtime.streamrecord.LatencyMarker;
+import org.apache.flink.streaming.runtime.streamrecord.StreamRecord;
+import org.apache.flink.streaming.runtime.streamstatus.StreamStatusProvider;
+import org.apache.flink.util.OutputTag;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import javax.annotation.Nullable;
+
+class ChainingOutput<T> implements WatermarkGaugeExposingOutput<StreamRecord<T>> {
+	private static final Logger LOG = LoggerFactory.getLogger(ChainingOutput.class);
+
+	protected final OneInputStreamOperator<T, ?> operator;
+	protected final Counter numRecordsIn;
+	protected final WatermarkGauge watermarkGauge = new WatermarkGauge();
+	protected final StreamStatusProvider streamStatusProvider;
+	@Nullable protected final OutputTag<T> outputTag;
+
+	public ChainingOutput(
+			OneInputStreamOperator<T, ?> operator,
+			StreamStatusProvider streamStatusProvider,
+			@Nullable OutputTag<T> outputTag) {
+		this.operator = operator;
+
+		{
+			Counter tmpNumRecordsIn;
+			try {
+				OperatorIOMetricGroup ioMetricGroup = ((OperatorMetricGroup) operator.getMetricGroup()).getIOMetricGroup();
+				tmpNumRecordsIn = ioMetricGroup.getNumRecordsInCounter();
+			} catch (Exception e) {
+				LOG.warn("An exception occurred during the metrics setup.", e);
+				tmpNumRecordsIn = new SimpleCounter();
+			}
+			numRecordsIn = tmpNumRecordsIn;
+		}
+
+		this.streamStatusProvider = streamStatusProvider;
+		this.outputTag = outputTag;
+	}
+
+	@Override
+	public void collect(StreamRecord<T> record) {
+		if (this.outputTag != null) {
+			// we are not responsible for emitting to the main output.
+			return;
+		}
+
+		pushToOperator(record);
+	}
+
+	@Override
+	public <X> void collect(OutputTag<X> outputTag, StreamRecord<X> record) {
+		if (this.outputTag == null || !this.outputTag.equals(outputTag)) {
+			// we are not responsible for emitting to the side-output specified by this
+			// OutputTag.
+			return;
+		}
+
+		pushToOperator(record);
+	}
+
+	protected <X> void pushToOperator(StreamRecord<X> record) {
+		try {
+			// we know that the given outputTag matches our OutputTag so the record
+			// must be of the type that our operator expects.
+			@SuppressWarnings("unchecked")
+			StreamRecord<T> castRecord = (StreamRecord<T>) record;
+
+			numRecordsIn.inc();
+			operator.setKeyContextElement1(castRecord);
+			operator.processElement(castRecord);
+		}
+		catch (Exception e) {
+			throw new ExceptionInChainedOperatorException(e);
+		}
+	}
+
+	@Override
+	public void emitWatermark(Watermark mark) {
+		try {
+			watermarkGauge.setCurrentWatermark(mark.getTimestamp());
+			if (streamStatusProvider.getStreamStatus().isActive()) {
+				operator.processWatermark(mark);
+			}
+		}
+		catch (Exception e) {
+			throw new ExceptionInChainedOperatorException(e);
+		}
+	}
+
+	@Override
+	public void emitLatencyMarker(LatencyMarker latencyMarker) {
+		try {
+			operator.processLatencyMarker(latencyMarker);
+		}
+		catch (Exception e) {
+			throw new ExceptionInChainedOperatorException(e);
+		}
+	}
+
+	@Override
+	public void close() {
+		try {
+			operator.close();
+		}
+		catch (Exception e) {
+			throw new ExceptionInChainedOperatorException(e);
+		}
+	}
+
+	@Override
+	public Gauge<Long> getWatermarkGauge() {
+		return watermarkGauge;
+	}
+}
